Admin->Manufacturers unable to delete manufacturers

+++ admin/applications/products/actions/deleteManufacturer.php
@@ -1,17 +1,18 @@
 <?php
 	$Manufacturer = Doctrine_Core::getTable('Manufacturers')->find($_GET['mID']);
 	if ($Manufacturer){
-		if (isset($_POST['delete_image']) && $_POST['delete_image'] == '1'){
+		if (isset($_POST['delete_image']) && $_POST['delete_image'] == '1' && strlen($Manufacturer->manufacturers_image) > 0){
 			$image_location = sysConfig::getDirFsCatalog() . 'images/' . $Manufacturer->manufacturers_image;
 
-			if (file_exists($image_location)) @unlink($image_location);
+			if (file_exists($image_location) && !is_dir($image_location)) @unlink($image_location);
 		}
-
-		$Products = $Manufacturer->Products;
-		if (isset($_POST['delete_products']) && $_POST['delete_products'] == '1'){
-			$Products->delete();
-		}else{
-			$Products->manufacturers_id = '';
+		if(count($Manufacturer->Products) > 0){
+			$Products = $Manufacturer->Products;
+			if (isset($_POST['delete_products']) && $_POST['delete_products'] == '1'){
+				$Products->delete();
+			}else{
+				$Products->manufacturers_id = '';
+			}
 		}
 		
 		$Manufacturer->delete();
